+++ source/common/upstream/cluster_manager_impl.h
@@ -22,6 +22,7 @@
 
 #include "common/config/grpc_mux_impl.h"
 #include "common/http/async_client_impl.h"
+#include "common/upstream/eds_subscription_factory.h"
 #include "common/upstream/load_stats_reporter.h"
 #include "common/upstream/upstream_impl.h"
 
@@ -39,9 +40,10 @@ public:
                             Ssl::ContextManager& ssl_context_manager,
                             Event::Dispatcher& main_thread_dispatcher,
                             const LocalInfo::LocalInfo& local_info,
-                            Secret::SecretManager& secret_manager, Api::Api& api)
-      : main_thread_dispatcher_(main_thread_dispatcher), api_(api), runtime_(runtime),
-        stats_(stats), tls_(tls), random_(random), dns_resolver_(dns_resolver),
+                            Secret::SecretManager& secret_manager, Api::Api& api,
+                            Http::Context& http_context)
+      : main_thread_dispatcher_(main_thread_dispatcher), api_(api), http_context_(http_context),
+        runtime_(runtime), stats_(stats), tls_(tls), random_(random), dns_resolver_(dns_resolver),
         ssl_context_manager_(ssl_context_manager), local_info_(local_info),
         secret_manager_(secret_manager) {}
 
@@ -76,6 +78,7 @@ public:
 protected:
   Event::Dispatcher& main_thread_dispatcher_;
   Api::Api& api_;
+  Http::Context& http_context_;
 
 private:
   Runtime::Loader& runtime_;
@@ -86,6 +89,7 @@ private:
   Ssl::ContextManager& ssl_context_manager_;
   const LocalInfo::LocalInfo& local_info_;
   Secret::SecretManager& secret_manager_;
+  Upstream::EdsSubscriptionFactory eds_subscription_factory_;
 };
 
 /**
@@ -172,8 +176,8 @@ public:
                      ThreadLocal::Instance& tls, Runtime::Loader& runtime,
                      Runtime::RandomGenerator& random, const LocalInfo::LocalInfo& local_info,
                      AccessLog::AccessLogManager& log_manager,
-                     Event::Dispatcher& main_thread_dispatcher, Server::Admin& admin,
-                     Api::Api& api);
+                     Event::Dispatcher& main_thread_dispatcher, Server::Admin& admin, Api::Api& api,
+                     Http::Context& http_context);
 
   // Upstream::ClusterManager
   bool addOrUpdateCluster(const envoy::api::v2::Cluster& cluster,
@@ -449,6 +453,7 @@ private:
   TimeSource& time_source_;
   ClusterUpdatesMap updates_map_;
   Event::Dispatcher& dispatcher_;
+  Http::Context& http_context_;
 };
 
 } // namespace Upstream