+++ source/extensions/filters/http/grpc_json_transcoder/json_transcoder_filter.cc
@@ -245,7 +245,9 @@ Http::FilterHeadersStatus JsonTranscoderFilter::decodeHeaders(Http::HeaderMap& h
     if (!request_status.ok()) {
       ENVOY_LOG(debug, "Transcoding request error {}", request_status.ToString());
       error_ = true;
-      decoder_callbacks_->sendLocalReply(Http::Code::BadRequest, request_status.error_message(),
+      decoder_callbacks_->sendLocalReply(Http::Code::BadRequest,
+                                         absl::string_view(request_status.error_message().data(),
+                                                           request_status.error_message().size()),
                                          nullptr, absl::nullopt);
 
       return Http::FilterHeadersStatus::StopIteration;
@@ -281,7 +283,9 @@ Http::FilterDataStatus JsonTranscoderFilter::decodeData(Buffer::Instance& data, 
   if (!request_status.ok()) {
     ENVOY_LOG(debug, "Transcoding request error {}", request_status.ToString());
     error_ = true;
-    decoder_callbacks_->sendLocalReply(Http::Code::BadRequest, request_status.error_message(),
+    decoder_callbacks_->sendLocalReply(Http::Code::BadRequest,
+                                       absl::string_view(request_status.error_message().data(),
+                                                         request_status.error_message().size()),
                                        nullptr, absl::nullopt);
 
     return Http::FilterDataStatus::StopIterationNoBuffer;