+++ source/extensions/filters/http/ratelimit/ratelimit.h
@@ -6,6 +6,7 @@
 #include <vector>
 
 #include "envoy/config/filter/http/rate_limit/v2/rate_limit.pb.h"
+#include "envoy/http/context.h"
 #include "envoy/http/filter.h"
 #include "envoy/local_info/local_info.h"
 #include "envoy/ratelimit/ratelimit.h"
@@ -15,6 +16,8 @@
 
 #include "common/common/assert.h"
 #include "common/http/header_map_impl.h"
+
+#include "extensions/filters/common/ratelimit/ratelimit.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -33,7 +36,7 @@ class FilterConfig {
 public:
   FilterConfig(const envoy::config::filter::http::rate_limit::v2::RateLimit& config,
                const LocalInfo::LocalInfo& local_info, Stats::Scope& scope,
-               Runtime::Loader& runtime)
+               Runtime::Loader& runtime, Http::Context& http_context)
       : domain_(config.domain()), stage_(static_cast<uint64_t>(config.stage())),
         request_type_(config.request_type().empty() ? stringToType("both")
                                                     : stringToType(config.request_type())),
@@ -42,7 +45,8 @@ public:
         rate_limited_grpc_status_(
             config.rate_limited_as_resource_exhausted()
                 ? absl::make_optional(Grpc::Status::GrpcStatus::ResourceExhausted)
-                : absl::nullopt) {}
+                : absl::nullopt),
+        http_context_(http_context) {}
   const std::string& domain() const { return domain_; }
   const LocalInfo::LocalInfo& localInfo() const { return local_info_; }
   uint64_t stage() const { return stage_; }
@@ -53,6 +57,7 @@ public:
   const absl::optional<Grpc::Status::GrpcStatus> rateLimitedGrpcStatus() const {
     return rate_limited_grpc_status_;
   }
+  Http::Context& httpContext() { return http_context_; }
 
 private:
   static FilterRequestType stringToType(const std::string& request_type) {
@@ -74,6 +79,7 @@ private:
   Runtime::Loader& runtime_;
   const bool failure_mode_deny_;
   const absl::optional<Grpc::Status::GrpcStatus> rate_limited_grpc_status_;
+  Http::Context& http_context_;
 };
 
 typedef std::shared_ptr<FilterConfig> FilterConfigSharedPtr;
@@ -82,9 +88,9 @@ typedef std::shared_ptr<FilterConfig> FilterConfigSharedPtr;
  * HTTP rate limit filter. Depending on the route configuration, this filter calls the global
  * rate limiting service before allowing further filter iteration.
  */
-class Filter : public Http::StreamFilter, public RateLimit::RequestCallbacks {
+class Filter : public Http::StreamFilter, public Filters::Common::RateLimit::RequestCallbacks {
 public:
-  Filter(FilterConfigSharedPtr config, RateLimit::ClientPtr&& client)
+  Filter(FilterConfigSharedPtr config, Filters::Common::RateLimit::ClientPtr&& client)
       : config_(config), client_(std::move(client)) {}
 
   // Http::StreamFilterBase
@@ -104,20 +110,22 @@ public:
   void setEncoderFilterCallbacks(Http::StreamEncoderFilterCallbacks& callbacks) override;
 
   // RateLimit::RequestCallbacks
-  void complete(RateLimit::LimitStatus status, Http::HeaderMapPtr&& headers) override;
+  void complete(Filters::Common::RateLimit::LimitStatus status,
+                Http::HeaderMapPtr&& headers) override;
 
 private:
   void initiateCall(const Http::HeaderMap& headers);
   void populateRateLimitDescriptors(const Router::RateLimitPolicy& rate_limit_policy,
-                                    std::vector<RateLimit::Descriptor>& descriptors,
+                                    std::vector<Envoy::RateLimit::Descriptor>& descriptors,
                                     const Router::RouteEntry* route_entry,
                                     const Http::HeaderMap& headers) const;
   void addHeaders(Http::HeaderMap& headers);
+  Http::Context& httpContext() { return config_->httpContext(); }
 
   enum class State { NotStarted, Calling, Complete, Responded };
 
   FilterConfigSharedPtr config_;
-  RateLimit::ClientPtr client_;
+  Filters::Common::RateLimit::ClientPtr client_;
   Http::StreamDecoderFilterCallbacks* callbacks_{};
   State state_{State::NotStarted};
   Upstream::ClusterInfoConstSharedPtr cluster_;