+++ test/extensions/filters/listener/tls_inspector/tls_inspector_test.cc
@@ -31,7 +31,6 @@ public:
   TlsInspectorTest() : cfg_(std::make_shared<Config>(store_)) {}
 
   void init() {
-    timer_ = new NiceMock<Event::MockTimer>(&dispatcher_);
     filter_ = std::make_unique<Filter>(cfg_);
     EXPECT_CALL(cb_, socket()).WillRepeatedly(ReturnRef(socket_));
     EXPECT_CALL(cb_, dispatcher()).WillRepeatedly(ReturnRef(dispatcher_));
@@ -54,7 +53,6 @@ public:
   Network::MockConnectionSocket socket_;
   NiceMock<Event::MockDispatcher> dispatcher_;
   Event::FileReadyCb file_event_callback_;
-  Event::MockTimer* timer_{};
 };
 
 // Test that an exception is thrown for an invalid value for max_client_hello_size
@@ -69,14 +67,6 @@ TEST_F(TlsInspectorTest, ConnectionClosed) {
   EXPECT_CALL(cb_, continueFilterChain(false));
   file_event_callback_(Event::FileReadyType::Closed);
   EXPECT_EQ(1, cfg_->stats().connection_closed_.value());
-}
-
-// Test that the filter detects timeout and terminates.
-TEST_F(TlsInspectorTest, Timeout) {
-  init();
-  EXPECT_CALL(cb_, continueFilterChain(false));
-  timer_->callback_();
-  EXPECT_EQ(1, cfg_->stats().read_timeout_.value());
 }
 
 // Test that the filter detects detects read errors.