+++ test/test_common/utility.h
@@ -16,6 +16,7 @@
 #include "envoy/thread/thread.h"
 
 #include "common/buffer/buffer_impl.h"
+#include "common/common/block_memory_hash_set.h"
 #include "common/common/c_smart_ptr.h"
 #include "common/common/thread.h"
 #include "common/http/header_map_impl.h"
@@ -412,52 +413,40 @@ makeHeaderMap(const std::initializer_list<std::pair<std::string, std::string>>& 
 namespace Stats {
 
 /**
- * This is a heap test allocator that works similar to how the shared memory allocator works in
- * terms of reference counting, etc.
+ * Implements a RawStatDataAllocator using a contiguous block of heap-allocated
+ * memory, but is otherwise identical to the shared memory allocator in terms of
+ * reference counting, data structures, etc.
  */
 class TestAllocator : public RawStatDataAllocator {
 public:
-  TestAllocator(const StatsOptions& stats_options) : stats_options_(stats_options) {}
-  ~TestAllocator() { EXPECT_TRUE(stats_.empty()); }
-
-  RawStatData* alloc(absl::string_view name) override {
-    CSmartPtr<RawStatData, freeAdapter>& stat_ref = stats_[std::string(name)];
-    if (!stat_ref) {
-      stat_ref.reset(static_cast<RawStatData*>(
-          ::calloc(RawStatData::structSizeWithOptions(stats_options_), 1)));
-      stat_ref->initialize(name, stats_options_);
-    } else {
-      stat_ref->ref_count_++;
+  struct TestBlockMemoryHashSetOptions : public BlockMemoryHashSetOptions {
+    TestBlockMemoryHashSetOptions() {
+      capacity = 200;
+      num_slots = 131;
     }
+  };
 
-    return stat_ref.get();
-  }
-
-  void free(RawStatData& data) override {
-    if (--data.ref_count_ > 0) {
-      return;
-    }
-
-    if (stats_.erase(std::string(data.name_)) == 0) {
-      FAIL();
-    }
-  }
+  explicit TestAllocator(const StatsOptions& stats_options)
+      : RawStatDataAllocator(mutex_, hash_set_, stats_options),
+        block_memory_(std::make_unique<uint8_t[]>(
+            RawStatDataSet::numBytes(block_hash_options_, stats_options))),
+        hash_set_(block_hash_options_, true /* init */, block_memory_.get(), stats_options) {}
+  ~TestAllocator() { EXPECT_EQ(0, hash_set_.size()); }
 
 private:
-  static void freeAdapter(RawStatData* data) { ::free(data); }
-  std::unordered_map<std::string, CSmartPtr<RawStatData, freeAdapter>> stats_;
-  const StatsOptions& stats_options_;
+  Thread::MutexBasicLockable mutex_;
+  TestBlockMemoryHashSetOptions block_hash_options_;
+  std::unique_ptr<uint8_t[]> block_memory_;
+  RawStatDataSet hash_set_;
 };
 
-class MockedTestAllocator : public RawStatDataAllocator {
+class MockedTestAllocator : public TestAllocator {
 public:
   MockedTestAllocator(const StatsOptions& stats_options);
   virtual ~MockedTestAllocator();
 
   MOCK_METHOD1(alloc, RawStatData*(absl::string_view name));
   MOCK_METHOD1(free, void(RawStatData& data));
-
-  TestAllocator alloc_;
 };
 
 } // namespace Stats